Refactor for provide/inject pattern
