+++ app/src/modules/content/routes/item.vue
@@ -24,6 +24,7 @@ import { useRouter, useRoute } from 'vue-router';
 import ContentNavigation from '../components/navigation.vue';
 import VersionMenu from '../components/version-menu.vue';
 import ContentNotFound from './not-found.vue';
+import { useFlows } from '@/composables/use-flows';
 
 interface Props {
 	collection: string;
@@ -500,6 +501,33 @@ function useCollectionRoute() {
 
 	return { collectionRoute };
 }
+
+function handleRunManualFlow(flowId: string, isActionDisabled = false) {
+	runManualFlow(flowId, isActionDisabled, refresh);
+}
+
+const {
+	confirmButtonCTA,
+	confirmDialogDetails,
+	confirmRunFlow,
+	confirmUnsavedChanges,
+	confirmValues,
+	displayCustomConfirmDialog,
+	displayUnsavedChangesDialog,
+	isConfirmButtonDisabled,
+	provideUseFlows,
+	resetConfirm,
+	runManualFlow,
+} = useFlows({
+	collection,
+	primaryKey: computed(() => actualPrimaryKey.value ?? undefined),
+	selection: ref([]),
+	location: ref('item'),
+	hasEdits,
+	onRefreshCallback: refresh,
+});
+
+provideUseFlows();
 </script>
 
 <template>
@@ -727,6 +755,60 @@ function useCollectionRoute() {
 					</v-menu>
 				</template>
 			</v-button>
+
+			<v-dialog
+				:model-value="displayUnsavedChangesDialog"
+				keep-behind
+				@esc="resetConfirm"
+				@apply="confirmUnsavedChanges(confirmRunFlow!)"
+			>
+				<v-card>
+					<v-card-title>{{ t('unsaved_changes') }}</v-card-title>
+					<v-card-text>{{ t('run_flow_on_current_edited_confirm') }}</v-card-text>
+
+					<v-card-actions>
+						<v-button secondary @click="resetConfirm">
+							{{ t('cancel') }}
+						</v-button>
+						<v-button :disabled="isConfirmButtonDisabled" @click="confirmUnsavedChanges(confirmRunFlow!)">
+							{{ confirmButtonCTA }}
+						</v-button>
+					</v-card-actions>
+				</v-card>
+			</v-dialog>
+
+			<v-dialog
+				:model-value="displayCustomConfirmDialog"
+				keep-behind
+				@esc="resetConfirm"
+				@apply="handleRunManualFlow(confirmRunFlow!, isConfirmButtonDisabled)"
+			>
+				<v-card>
+					<v-card-title>{{ confirmDialogDetails!.description ?? t('run_flow_confirm') }}</v-card-title>
+					<v-card-text class="confirm-form">
+						<v-form
+							v-if="confirmDialogDetails!.fields && confirmDialogDetails!.fields.length > 0"
+							:fields="confirmDialogDetails!.fields"
+							:model-value="confirmValues"
+							autofocus
+							primary-key="+"
+							@update:model-value="confirmValues = $event"
+						/>
+					</v-card-text>
+
+					<v-card-actions>
+						<v-button secondary @click="resetConfirm">
+							{{ t('cancel') }}
+						</v-button>
+						<v-button
+							:disabled="isConfirmButtonDisabled"
+							@click="handleRunManualFlow(confirmRunFlow!, isConfirmButtonDisabled)"
+						>
+							{{ confirmButtonCTA }}
+						</v-button>
+					</v-card-actions>
+				</v-card>
+			</v-dialog>
 		</template>
 
 		<template #navigation>
