+++ app/src/views/private/components/flow-sidebar-detail.vue
@@ -1,57 +1,18 @@
 <script setup lang="ts">
-import { useFlows } from '@/composables/use-flows';
-import { toRefs } from 'vue';
+import { inject } from 'vue';
 import { useI18n } from 'vue-i18n';
 import TriggerFlow from './trigger-flow.vue';
 
-const props = withDefaults(
-	defineProps<{
-		collection: string;
-		primaryKey?: string | number;
-		selection?: (number | string)[];
-		location: 'collection' | 'item';
-		hasEdits?: boolean;
-	}>(),
-	{
-		primaryKey: undefined,
-		selection: () => [],
-		hasEdits: false,
-	},
-);
-
-const emit = defineEmits(['refresh']);
-
 const { t } = useI18n();
 
-const { collection, primaryKey, selection, location, hasEdits } = toRefs(props);
-
-const { manualFlows } = useFlows({
-	collection,
-	primaryKey,
-	selection,
-	location,
-	hasEdits,
-	onRefreshCallback: handleRefresh,
-});
-
-function handleRefresh() {
-	emit('refresh');
-}
+const { manualFlows } = inject('provide-use-flows') as any;
 </script>
 
 <template>
 	<sidebar-detail v-if="manualFlows.length > 0" icon="bolt" :title="t('flows')">
 		<div class="fields">
 			<div v-for="manualFlow in manualFlows" :key="manualFlow.id" class="field full">
-				<trigger-flow
-					:flow-id="manualFlow.id"
-					:collection="collection"
-					:primary-key="primaryKey"
-					:selection="selection"
-					:location="location"
-					:has-edits="hasEdits"
-					@refresh="handleRefresh"
-				/>
+				<trigger-flow :flow-id="manualFlow.id" />
 			</div>
 		</div>
 	</sidebar-detail>
